+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/database/AlbumDatabase.kt
@@ -2,7 +2,7 @@ package com.igorwojda.showcase.feature.album.data.database
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
-import com.igorwojda.showcase.feature.album.data.network.model.AlbumEntity
+import com.igorwojda.showcase.feature.album.data.database.model.AlbumEntity
 
 @Database(entities = [AlbumEntity::class], version = 1, exportSchema = false)
 abstract class AlbumDatabase : RoomDatabase() {
