+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/network/enum/AlbumDataImageSize.kt
@@ -1,6 +1,6 @@
 package com.igorwojda.showcase.feature.album.data.network.enum
 
-import com.igorwojda.showcase.feature.album.data.network.model.AlbumImageSizeEntity
+import com.igorwojda.showcase.feature.album.data.database.model.AlbumImageSizeEntity
 import com.igorwojda.showcase.feature.album.domain.enum.AlbumDomainImageSize
 import com.squareup.moshi.Json
