+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/network/model/AlbumImageNetwork.kt
@@ -1,5 +1,6 @@
 package com.igorwojda.showcase.feature.album.data.network.model
 
+import com.igorwojda.showcase.feature.album.data.database.model.AlbumImageEntity
 import com.igorwojda.showcase.feature.album.data.network.enum.AlbumDataImageSize
 import com.igorwojda.showcase.feature.album.data.network.enum.toDomainEnum
 import com.igorwojda.showcase.feature.album.data.network.enum.toEntityEnum
