+++ feature_album/src/main/kotlin/com/igorwojda/showcase/feature/album/data/network/model/AlbumNetwork.kt
@@ -1,5 +1,6 @@
 package com.igorwojda.showcase.feature.album.data.network.model
 
+import com.igorwojda.showcase.feature.album.data.database.model.AlbumEntity
 import com.igorwojda.showcase.feature.album.data.network.enum.AlbumDataImageSize
 import com.igorwojda.showcase.feature.album.domain.model.Album
 import com.squareup.moshi.Json
